Improve messaing when receipt already found

+++ src/cli/subcommand/install.rs
@@ -94,6 +94,20 @@ impl CommandExecute for Install {
 
                 match existing_receipt {
                     Some(existing_receipt) => {
+                        if let Err(e) = existing_receipt.check_compatible() {
+                            eprintln!(
+                                "{}", 
+                                format!("
+                                    {e}\n\
+                                    \n\
+                                    Found existing plan in `{RECEIPT_LOCATION}` which was created by a version incompatible `nix-installer`.\n\
+                                    If you are trying to upgrade Nix, try running `sudo -i nix upgrade-nix` instead.
+                                    If you are trying to install Nix over an existing install (from an incompatible `nix-installer` install), try running `/nix/nix-installer uninstall` then try to install again.\n\
+                                    If you are using `nix-installer` in an automated curing process and seeing this message, consider pinning the version you use via https://github.com/DeterminateSystems/nix-installer#accessing-other-versions.\n\
+                                ").red()
+                            );
+                            return Ok(ExitCode::FAILURE)
+                        }
                         if existing_receipt.planner.typetag_name() != chosen_planner.typetag_name() {
                             eprintln!("{}", format!("Found existing plan in `{RECEIPT_LOCATION}` which used a different planner, try uninstalling the existing install with `{uninstall_command}`").red());
                             return Ok(ExitCode::FAILURE)
@@ -104,7 +118,7 @@ impl CommandExecute for Install {
                         }
                         eprintln!("{}", format!("Found existing plan in `{RECEIPT_LOCATION}`, with the same settings, already completed, try uninstalling (`{uninstall_command}`) and reinstalling if Nix isn't working").red());
                         return Ok(ExitCode::FAILURE)
-                    } ,
+                    },
                     None => {
                         let res = planner.plan().await;
                         match res {
